// src/common/gateways/cnab/enums.ts
export enum TipoDocumento {
  ISENTO = "0",
  CPF = "1",
  CNPJ = "2",
  PIS_PASEP = "3",
  OUTRO = "9",
}

export enum CodCamaraCentralizadora {
  TED = "018",
  CC = "000",
  DOC = "700",
  PIX = "009",
}

export enum TipoMovimento {
  INCLUSAO = "0",
  CONSULTA = "1",
  ESTORNO = "3",
  ALTERACAO = "5",
  LIQUIDACAO = "7",
  EXCLUSAO = "9",
}
// ...
export enum CodigoMovimento {
  INCLUSAO = "00",
  ALTERACAO = "05",
  ALTERACAO_AUTORIZA = "06",
  INCLUSAO_BLOQUEADO = "09",
  ALTERACAO_BLOQUEADO = "10",
  ALTERACAO_LIBERADO = "11",
  ALTERACAO_VALOR = "17",
  ALTERACAO_DATA_PAGAMENTO = "19",
  PAGAMENTAMENTO_DIRETO_FORNECEDOR = "23",
  MANUTENCAO_CARTEIRA = "25",
  RETIRADA_CARTEIRA = "27",
  ESTORNO_DEVOLUCAO = "33",
  ALEGACAO_PAGADOR = "40",
  INCLUSAO_CONTA_FAVORECIDO = "50",
  ALTERACAO_CONTA_FAVORECIDO = "51",
  BLOQUEIO_CONTA_FAVORECIDO = "52",
  DESBLOQUEIO_CONTA_FAVORECIDO = "53",
  EXCLUSAO_CONTA_FAVORECIDO = "54",
  INCLUSAO_FAVORECIDOS = "60",
  ALTERACAO_FAVORECIDOS = "61",
  BLOQUEIO_FAVORECIDOS = "62",
  DESBLOQUEIO_FAVORECIDOS = "63",
  ABERTURA_MASSIVA_CONTAS = "64",
  EXCLUSAO_FAVORECIDOS = "70",
  EXCLUSAO_MASSIVA_CONTAS = "71",
  EXCLUSAO_REGISTRO_DETALHE = "99",
}

export enum IdentificacaoFavorecidoChavePix {
  TELEFONE = "01",
  EMAIL = "02",
  CNPJ_CPF = "03",
  ALEATORIA = "04",
  DADOS_BANCARIOS = "05",
}

export enum TipoServico {
  COBRANCA = "01",
  BLOQUETO_ELETRONICO = "03",
  CONCILIACAO_BANCARIA = "04",
  DEBITOS = "05",
  CUSTODIA_CHEQUES = "06",
  GESTAO_CAIXA = "07",
  CONSULTA_MARGEM = "08",
  AVERBACAO_CONSIGNACAO = "09",
  PAGAMENTO_DIVIDENDOS = "10",
  MANUTENCAO_CONSIGNACAO = "11",
  CONSIGNACAO_PARCELAS = "12",
  GLOSA_CONSIGNACAO = "13",
  CONSULTA_TRIBUTOS_PAGAR = "14",
  PAGAMENTO_FORNECEDORES = "20",
  PAGAMENTO_CONTAS = "22",
  COMPROR = "25",
  COMPROR_ROTATIVO = "26",
  ALEGACAO_SACADO = "29",
  PAGAMENTO_SALARIOS = "30",
  PAGAMENTO_HONORARIOS = "32",
  PAGAMENTO_BOLSA_AUXILIO = "33",
  PAGAMNETO_PREBENDA = "34",
  VENDOR = "40",
  VENDOR_TERMO = "41",
  PAGAMENTO_SINISTROS = "50",
  PAGAMENTO_DESPESAS_VIAJANTE = "60",
  PAGAMENTO_AUTORIZADO = "70",
  PAGAMENTO_CREDENCIADOS = "75",
  PAGAMENTO_REMUNERACAO = "77",
  PAGAMENTO_REPRESENTANTES = "80",
  PAGAMENTO_BENEFICIOS = "90",
  PAGAMENTOS_DIVERSOS = "98",
  EXCLUSIVO_BRADESCO = "99",
}

export enum FormaLancamento {
  CREDITO_CC = "01", // transferência para contas do Bradesco
  CEDITO_ADM = "02",
  DOC_TED = "03",
  CARTAO_SALARIO = "04", // Somente para tipo serviço 30
  CREDITO_POUPANCA = "05",
  OP_DISPOSICAO = "10",
  /**
   * Pagamento de Contas e Tributos com Código de Barras
   */
  PGTO_CONTAS = "11",
  TRIB_DARF_NORMAL = "16",
  TRIB_GPS = "17",
  TRIB_DARF_SIMPLES = "18",
  TRIB_IPTU_PREFEITURA = "19",
  PGTO_AUTENTICACAO = "20",
  TRIB_DARJ = "21",
  TRIB_GARE_ICMS = "22",
  TRIB_GARE_DR = "23",
  TRIB_GARE_ITCMD = "24",
  TRIB_IPVA = "25",
  TRIB_LICENCIAMENTO = "26",
  TRIB_DPVAT = "27",
  LIQUIDACAO_TITULOS_BANCO = "30",
  PGTO_TITULOS_OUTRO_BANCO = "31",
  EXTRATO_CONTA_CORRENTE = "40",
  TED_OUTRA_TITULARIDADE = "41",
  TED_MESMA_TITULARIDADE = "43",
  TED_CONTA_INVESTIMENTO = "44",
  PIX_TRANSFERENCIA = "45",
  PIX_QR_CODE = "47",
  DEBITO_CONTA_CORRENTE = "50",
  EXTRATO_GESTAO_CAIXA = "70",
  DEPOSITO_JUDICIAL_CC = "71",
  DEPOSITO_JUDICIAL_CP = "72",
  EXTRATO_CONTA_INVESTIMENTO = "73",
  CADASTRO_FAVORECIDOS = "99",
}

export enum ComplementoTipoServico {
  CREDITO_CONTA = "01",
  PAGAMENTO_ALUGUEL_CONDOMINIO = "02",
  PAGAMENTO_DUPLICATA_TITULOS = "03",
  PAGAMENTO_DIVIDENDOS = "04",
  PAGAMENTO_MENSALIDADE_ESCOLAR = "05",
  PAGAMENTO_SALARIOS = "06",
  PAGAMENTO_FORNECEDORES = "07",
  OPERACOES_CAMBIOS_FUNDOS_BOLSA = "08",
  REPASSE_ARRECADACAO_TRIBUTOS = "09",
  TRANSFERENCIA_INTERNACIONAL_REAL = "10",
  DOC_POUPANCA = "11",
  DOC_DEPOSITO_JUDICIAL = "12",
  OUTROS = "13",
  PAGAMENTO_BOLSA_AUXILIO = "16",
  REMUNERACAO_COOPERADO = "17",
  PAGAMENTO_HONORARIOS = "18",
  PAGAMENTO_PREBENDA = "19",
}

export enum ComplementoTipoServicoTED {
  IMPOSTOS_TRIBUTOS_TAXAS = "00001",
  CONCESSAO_SERVICOS_PUBLICOS = "00002",
  PAGAMENTO_DIVIDENDOS = "00003",
  PAGAMENTO_SALARIOS = "00004",
  PAGAMENTO_FORNECEDORES = "00005",
  PAGAMENTO_HONORARIOS = "00006",
  ALUGUEL_TAXA_CONDOMINIO = "00007",
  DUPLICATAS_TITULOS = "00008",
  MENSALIDADE_ESCOLAR = "00009",
  CREDITO_CONTA = "00010",
  PAGAMENTO_CORRETORAS = "00011",
  BOLETO_BANCARIO_CARTORIO = "00012",
  TARIFAS_PRESTACAO_SERVICO_CONVENIO = "00013",
  REPASSE_VALOR_TITULO_LIQUIDADO_CARTORIO = "00014",
  LIQUIDACAO_FINANCEIRA_OPERADORA_CARTAO = "00015",
  OPERACOES_SEGURO_HABITACIONAL = "00018",
  OPERACOES_FUNDOS_CAIXA = "00019",
  TAXA_ADMINISTRACAO = "00023",
  ACORDO_EXECUCAO_JUDICIAL = "00027",
  LIQUIDACAO_EMPRESTIMO_CONSIGNADO = "00028",
  PAGAMENTO_BOLSA_AUXILIO = "00029",
  REMUNERACAO_COOPERADO = "00030",
  PAGAMENTO_PREBENDA = "00031",
  JUROS_SOBRE_CAPITAL = "00033",
  RENDIMENTOS_AMORTIZACAO = "00034",
  TAXA_SERVICO = "00035",
  JUROS_AMORTIZACAO_TITULO = "00037",
  ESTORNO_RESTITUICAO = "00038",
  PAGAMENTO_VALE_TRANSPORTE = "00039",
  SIMPLES_NACIONAL = "00040",
  REPASSE_VALORES_FUNDEB = "00041",
  REPASSE_VALORES_CONVENIO_CENTRALIZADO = "00042",
  PATROCINIO_INCENTIVO_FISCAL = "00043",
  DOACAO_INCENTIVO_FISCAL = "00044",
  FINANCIAMENTO_CONTA_CORRENTE = "00045",
  PAGAMENTO_RESCISAO_CONTRATO_TRABALHO = "00047",
  REEMBOLSO_DESPESA_ESTRUTURA_OPERACAO = "00050",
  COMPRA_MOEDA_ESTRANGEIRA = "00097",
  DEPOSITO_JUDICIAL = "00100",
  PENSAO_ALIMENTICIA = "00101",
  CESSAO_CREDITO_LIQUIDACAO_PRINCIPAL_AQUISICAO = "00103",
  CESSAO_CREDITO_LIQUIDACAO_PRINCIPAL_FIDC = "00104",
  CESSAO_CREDITO_REPASSE_CONTRATO_FLUXO_CAIXA = "00107",
  CESSAO_CREDITO_REPASSE_FLUXO_CAIXA_RECEBIVEL = "00108",
  CESSAO_CREDITO_AJUSTES_DIVERSOS = "00109",
  TRANSFERENCIA_CONTAS_MESMO_TITULAR = "00110",
  CREDITO_INVESTIMENTO_CLIENTE = "00111",
  DEBITO_INVESTIMENTO_CLIENTE = "00112",
  PAGAMENTO_OPERACAO_CREDITO_CLIENTE = "00113",
  RESGATE_INVESTIMENTO_CLIENTE = "00114",
  APLICACAO_FINANCEIRA_CLIENTE_REMETENTE = "00117",
  CESSAO_CREDITO_LIQUIDACAO_RECOMPRA = "00123",
  CESSAO_CREDITO_LIQUIDACAO_RECOMPRA_FIDC = "00124",
  FGCOOP_RECOLHIMENTO = "00131",
  FGCOOP_DEVOLUCAO_RECOLHIMENTO = "00132",
  TRANSFERENCIA_INTERNACIONAL_REAIS = "00200",
  AJUSTE_POSICAO_MERCADO_FUTURO = "00201",
  REPASSE_VALORES_BNDES = "00202",
  LIQUIDACAO_COMPROMISSO_BNDES = "00203",
  OPERACOES_COMPRA_VENDA_ACOES = "00204",
  CONTRATO_REFERENCIA_ACOES_INDICE = "00205",
  OPERACAO_CAMBIO = "00206",
  OPERACAO_MERCADO_RENDA_FIXA_VARIAVEL = "00207",
  OPERACAO_CAMBIO_MERCADO_INTERBANCARIO = "00208",
  PAGAMENTO_OPERACAO_IDENTIFICACAO_DESTINATARIO = "00209",
  RESTITUICAO_IMPOSTO_RENDA = "00300",
  ORDEM_BANCARIA_TESOURO = "00301",
  FINANCIAMENTO_PAGAMENTO_MULTA_BACEN = "00302",
  FINANCIAMENTO_PAGAMENTO_RESTITUICAO_TRIBUTO_RFB = "00303",
  RESTITUICAO_PREMIOS_SEGUROS = "00500",
  PAGAMENTO_INDENIZACAO_SINISTRO_SEGURO = "00501",
  PAGAMENTO_PREMIO_COSSEGURO = "00502",
  PAGAMENTO_INDENIZACAO_SINISTRO_COSSEGURO = "00504",
  PAGAMENTO_PREMIO_RESSEGURO = "00505",
  PAGAMENTO_INDENIZACAO_SINISTRO_RESSEGURO = "00507",
  RESTITUICAO_INDENIZACAO_SINISTRO_RESSEGURO = "00508",
  PAGAMENTO_DESPESAS_SINISTROS = "00509",
  PAGAMENTO_INSPECOES_VISTORIA_PREVIA = "00510",
  PAGAMENTO_RESGATE_TITULO_CAPITALIZACAO = "00511",
  PAGAMENTO_SORTEIO_TITULO_CAPITALIZACAO = "00512",
  PAGAMENTO_DEVOLUCAO_MENSALIDADE_TITULO_CAPITALIZACAO = "00513",
  RESTITUICAO_CONTRIBUICAO_PLANO_PREVIDENCIA = "00514",
  PAGAMENTO_BENEFICIO_PREVIDENCIA_PECULIO = "00515",
  PAGAMENTO_BENEFICIO_PREVIDENCIA_PENSAO = "00516",
  PAGAMENTO_BENEFICIO_PREVIDENCIA_APOSENTADORIA = "00517",
  PAGAMENTO_RESGATE_PREVIDENCIARIO = "00518",
  PAGAMENTO_COMISSAO_CORRETAGEM = "00519",
  PAGAMENTO_TRANSFERENCIA_PORTABILIDADE = "00520",
}